MNT: Remove circular dependency which stopped `_shared.utils` from being imported.
